==> notebooks/1_se_data_preprocessing.ipynb <==
This script shows how our data preprocessing steps that were done to generate and store the pre-processed data as pickle files to be used in fine-tuning our models as well as to visualize our raw data.

The raw data was downloaded from https://www.lib.umich.edu/database/link/10387 and can be found in ../data/social_explorer_raw

In that directory you will see several other datasets we were aspiring to include, but were not able to get to in time

import pandas as pd

Crime Data: 

Initially there were many fields representing different slices of the big picture, we felt that trimming this down to just 1 field would be better for fine-tuning. Towards this end we used two of the existing fields to calculate total violent and property crimes as a percent of the total population.

crime_df = pd.read_csv('../data/social_explorer_raw/crime.csv')
crime_df = crime_df.drop(index=0)
crime_df = crime_df.dropna()
crime_df = crime_df.reset_index(drop=True)

crime_df['Total Violent and Property Crimes '] = crime_df['Total Violent and Property Crimes '].astype('float')
crime_df['Total Population (2014 est.)'] = crime_df['Total Population (2014 est.)'].astype('float')
crime_df['Total Violent and Property Crimes'+' %'] = (crime_df['Total Violent and Property Crimes '] / crime_df['Total Population (2014 est.)']) * 100


# crime_df = crime_df[['Qualifying Name', 'Total Violent and Property Crimes %']].round(2)
crime_df = crime_df[['FIPS', 'Qualifying Name','Total Violent and Property Crimes %']].round(2)
crime_df.to_pickle('../data/social_explorer_processed/crime_df.pkl')
crime_df.head(100)

Demographic data:

We shrink the demographic dataset to focus on population number of each ethnic group of the latest year(year 2020). We do not care about historical data as we are not interested in analyzing trends. We omit cols such as 2010 total population and ethic population of ethnic group in 2010.

Similar to crime data we calculated these as a percent of the total population.

demographic_df = pd.read_csv('../data/social_explorer_raw/Census2020_PL94_Redistricting_Data.csv')
demographic_df = demographic_df.drop(index=0)
demographic_df = demographic_df.reset_index(drop=True)

demographic_col_to_keep = ['FIPS', 'Qualifying Name','2020 Total Population','2020 Population Density', '2020 Housing Units',
       '2020 Occupied Units', '2020 Vacant Units',
       'White 2020', 'African American or Black 2020',
       'American Indian and Alaska Native 2020', 'Asian 2020',
       'Native Hawaiian and Other Pacific Islander 2020', 'Other 2020',
       'Two or More Races 2020',
       'Non-Hispanic 2020', 'Hispanic 2020']
demographic_df_filtered = demographic_df[demographic_col_to_keep]

columns_to_convert = [
       'White 2020', 'African American or Black 2020',
       'American Indian and Alaska Native 2020', 'Asian 2020',
       'Native Hawaiian and Other Pacific Islander 2020',
       'Other 2020', 'Two or More Races 2020',
       'Non-Hispanic 2020', 'Hispanic 2020'
]

demographic_df_filtered = demographic_df_filtered.dropna()
demographic_df_filtered['2020 Total Population'] = demographic_df_filtered['2020 Total Population'].astype('float')

for col in columns_to_convert:
    demographic_df_filtered[col] = demographic_df_filtered[col].astype('float')
    demographic_df_filtered[col+' %'] = (demographic_df_filtered[col] / demographic_df_filtered['2020 Total Population']) * 100

demographic_col_to_keep = ['FIPS','Qualifying Name','2020 Total Population','2020 Population Density',
       'White 2020 %', 'African American or Black 2020 %',
       'American Indian and Alaska Native 2020 %', 'Asian 2020 %',
       'Native Hawaiian and Other Pacific Islander 2020 %', 'Other 2020 %',
       'Two or More Races 2020 %', 'Hispanic 2020 %']
demographic_df = demographic_df_filtered[demographic_col_to_keep].round(2)
demographic_df.to_pickle('../data/social_explorer_processed/demographic_df.pkl')
demographic_df.head(100)

Housing Data:

The latest year of housing data available on Social Explorer was year 2020. This dataset include property rental data such as fair market rent.

Needed to remove the last 5 characters from FIPS code

housing_df = pd.read_csv('../data/social_explorer_raw/Income_limits_and_fair_market_rent_2020.csv')
housing_df = housing_df.drop(index=0)
housing_df = housing_df.reset_index(drop=True)

housing_col_to_keep = ['FIPS', 'Qualifying Name', 'Median Family Income (5-year ACS)', 'Fair Market Rent for One Bedroom',
       'Fair Market Rent for Two Bedrooms',
       'Fair Market Rent for Three Bedrooms',
       'Fair Market Rent for Four Bedrooms']
housing_df = housing_df[housing_col_to_keep]
housing_df['FIPS'] = housing_df['FIPS'].str[:-5]
housing_df.to_pickle('../data/social_explorer_processed/housing_df.pkl')
housing_df.head(100)

Environment Data:

Similar to the previous dataframes we re-calculated these area metrics as percent of total land.

environment_df = pd.read_csv('../data/social_explorer_raw/Environmental_summaries_2011.csv')
environment_df = environment_df.drop(index=0)
environment_df = environment_df.reset_index(drop=True)

col_to_keep = ['FIPS', 'Qualifying Name', 'Total Area of Land Coverage', 'Area Of Land Cover Class 11, Open Water'
       , 'Area Of Land Cover Class 12, Perennial Ice/Snow', 'Area Of Land Cover Class 21, Developed, Open Space', 'Area Of Land Cover Class 22, Developed, Low Intensity', 'Area Of Land Cover Class 23, Developed, Medium Intensity', 'Area Of Land Cover Class 24, Developed, High Intensity', 'Area Of Land Cover Class 31, Barren Land (Rock/Sand/Clay)','Area Of Land Cover Class 41, Deciduous Forest','Area Of Land Cover Class 42, Evergreen Forest','Area Of Land Cover Class 43, Mixed Forest', 'Area Of Land Cover Class 52, Shrub/Scrub', 'Area Of Land Cover Class 71, Grassland/Herbaceous', 'Area Of Land Cover Class 81, Pasture/Hay', 'Area Of Land Cover Class 82, Cultivated Crops', 'Area Of Land Cover Class 90, Woody Wetlands', 'Area Of Land Cover Class 95, Emergent Herbaceous Wetlands']

environment_df = environment_df[col_to_keep]

environment_df = environment_df.fillna(0)
environment_df['Total Area of Land Coverage'] = environment_df['Total Area of Land Coverage'].astype('float')

columns_to_convert = ['Area Of Land Cover Class 11, Open Water'
       , 'Area Of Land Cover Class 12, Perennial Ice/Snow', 'Area Of Land Cover Class 21, Developed, Open Space', 'Area Of Land Cover Class 22, Developed, Low Intensity', 'Area Of Land Cover Class 23, Developed, Medium Intensity', 'Area Of Land Cover Class 24, Developed, High Intensity', 'Area Of Land Cover Class 31, Barren Land (Rock/Sand/Clay)','Area Of Land Cover Class 41, Deciduous Forest','Area Of Land Cover Class 42, Evergreen Forest','Area Of Land Cover Class 43, Mixed Forest', 'Area Of Land Cover Class 52, Shrub/Scrub', 'Area Of Land Cover Class 71, Grassland/Herbaceous', 'Area Of Land Cover Class 81, Pasture/Hay', 'Area Of Land Cover Class 82, Cultivated Crops', 'Area Of Land Cover Class 90, Woody Wetlands', 'Area Of Land Cover Class 95, Emergent Herbaceous Wetlands']

for col in columns_to_convert:
    environment_df[col] = environment_df[col].astype('float')
    environment_df[col+' %'] = (environment_df[col] / environment_df['Total Area of Land Coverage']) * 100

col_to_keep = ['FIPS', 'Qualifying Name', 'Total Area of Land Coverage', 'Area Of Land Cover Class 11, Open Water %'
       , 'Area Of Land Cover Class 12, Perennial Ice/Snow %', 'Area Of Land Cover Class 21, Developed, Open Space %', 'Area Of Land Cover Class 22, Developed, Low Intensity %', 'Area Of Land Cover Class 23, Developed, Medium Intensity %', 'Area Of Land Cover Class 24, Developed, High Intensity %', 'Area Of Land Cover Class 31, Barren Land (Rock/Sand/Clay) %','Area Of Land Cover Class 41, Deciduous Forest %','Area Of Land Cover Class 42, Evergreen Forest %','Area Of Land Cover Class 43, Mixed Forest %', 'Area Of Land Cover Class 52, Shrub/Scrub %', 'Area Of Land Cover Class 71, Grassland/Herbaceous %', 'Area Of Land Cover Class 81, Pasture/Hay %', 'Area Of Land Cover Class 82, Cultivated Crops %', 'Area Of Land Cover Class 90, Woody Wetlands %', 'Area Of Land Cover Class 95, Emergent Herbaceous Wetlands %']    

environment_df = environment_df[col_to_keep].round(2)

shortened_columns = {
    'Total Area of Land Coverage': 'Total Area',
    'Area Of Land Cover Class 11, Open Water %': 'Open Water %',
    'Area Of Land Cover Class 12, Perennial Ice/Snow %': 'Ice/Snow %',
    'Area Of Land Cover Class 21, Developed, Open Space %': 'Developed Open Space %',
    'Area Of Land Cover Class 22, Developed, Low Intensity %': 'Developed Low Intensity %',
    'Area Of Land Cover Class 23, Developed, Medium Intensity %': 'Developed Medium Intensity %',
    'Area Of Land Cover Class 24, Developed, High Intensity %': 'Developed High Intensity %',
    'Area Of Land Cover Class 31, Barren Land (Rock/Sand/Clay) %': 'Barren Land %',
    'Area Of Land Cover Class 41, Deciduous Forest %': 'Deciduous Forest %',
    'Area Of Land Cover Class 42, Evergreen Forest %': 'Evergreen Forest %',
    'Area Of Land Cover Class 43, Mixed Forest %': 'Mixed Forest %',
    'Area Of Land Cover Class 52, Shrub/Scrub %': 'Shrub/Scrub %',
    'Area Of Land Cover Class 71, Grassland/Herbaceous %': 'Grassland/Herbaceous %',
    'Area Of Land Cover Class 81, Pasture/Hay %': 'Pasture/Hay %',
    'Area Of Land Cover Class 82, Cultivated Crops %': 'Cultivated Crops %',
    'Area Of Land Cover Class 90, Woody Wetlands %': 'Woody Wetlands %',
    'Area Of Land Cover Class 95, Emergent Herbaceous Wetlands %': 'Emergent Herbaceous Wetlands %'
}

environment_df.rename(columns=shortened_columns, inplace=True)
environment_df = environment_df.fillna(0)
environment_df.to_pickle('../data/social_explorer_processed/environment_df.pkl')
environment_df.head(100)

The following cell is commented out and was just used for spot checking results of various data, we decided to leave it as it may be useful to others exploring this work in the future!

# demographic_df_filtered = demographic_df[demographic_df['Qualifying Name'] == 'Santa Clara']
filtered_df_partial = environment_df[environment_df['Qualifying Name'].str.contains('Orange County')]
filtered_df_partial.head()
