==> notebooks/4_sagemaker_upload_train_data.ipynb <==
# Upload training data to AWS S3

We will be using Sagemaker Jumpstart to train the models - this requires the training data formatted in a specific way & uploaded to an S3 bucket

import pandas as pd

df = pd.read_pickle('../data/social_explorer/combined_df.pkl')

pd.options.display.max_rows = 4000

df.sample(5)['Combined'].values

import re

def split_combined(text):
    pattern = r"\[INST\](.*?)\[/INST\](.*)"
    match = re.match(pattern, text, re.DOTALL)
    
    if match:
        prompt = match.group(1).strip()
        answer = match.group(2).strip()
        return prompt, answer
    else:
        return None, None

df[['prompt', 'answer']] = df['Combined'].apply(lambda x: pd.Series(split_combined(x)))

print(df[['prompt', 'answer']].head())


def create_dialog_row(row):
    return {
        "dialog": [
            {"content": row['prompt'], "role": "user"},
            {"content": row['answer'], "role": "assistant"}
        ]
    }

df['dialog'] = df.apply(create_dialog_row, axis=1)

df = df.drop(columns=['Combined', 'prompt', 'answer'])

df.head()

df['dialog'].to_json('../data/social_explorer/formatted_dialog.jsonl', index=False, orient='records', lines=True)

import sagemaker

sagemaker_session = sagemaker.Session()

sagemaker_session.upload_data('../data/social_explorer/formatted_dialog.jsonl')
